[IMP] point_of_sale: Send ticket by email.
When you select a customer in the front end, you can send him an email with the receipt as an image.
If you decided to send the mail and you'd like to print the receipt anyway, you have to click on the
When you're offline, the order will be queued and the ticket printed. You'll receive an error message
When invoicing, the invoice will be in attachment of the email.

TASK-ID: 1922660

closes odoo/odoo#35812

Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>