[FIX] point_of_sale: total discount computation
When the total discount is computed on a receipt, it is computed based
on the the discount applied on lines, and the difference between the
pricelist price and the unit price. But there is an option on the
pricelist that say if you want to show or not the difference. Which is
not respected in this case because in total we always show all discount.

So we are now respecting the policy of the pricelist.

TASK-ID: 2058650