[IMP] stock: empty inventory report
Before this commit, if user tries to create a new record from a empty
grouped editable list in 'Inventory Report', he/she will get the form
view open which isn't editable.
To avoid that, we remove the group filter if there is no quant, so user
can create a new quant directly in the editable list.

Task #2056651

closes odoo/odoo#36221

Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>


Co-authored-by: jip-odoo <jip@odoo.com>