[FIX] point_of_sale: make partner creation work
Since changes made in the ORM, we cannot pass anymore string of integer
to many2One. We are now parsing the fields from the formular to
reprensent the integer as integer.

Since changes made on image fields, the "image" field name was not
correct so we renamed it in imgage_1920.

TASK-ID: 2058650

closes odoo/odoo#36219

Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>