[FIX] event: Allow to navigate from registration to event
This patch fixes a functional regression introduced in caf4d1e12305f347faebc38fa9e6d3069bbd70c9. Before that commit, users were able to navigate from a registration to the related event. Now it's impossible, and under some circumstances this confuses the user (i.e. when he comes to the form from an URL that another user shared to him, and there are no breadcrumbs to follow).

Now, users will be able to click in the event and go to its form.

closes odoo/odoo#37834

X-original-commit: eb46c13704bf3aa515fe3a92cc588b762c5e5164
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>